@@ -2,29 +2,36 @@ apply plugin: 'com.android.application'
 apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     buildToolsVersion "29.0.0"
-    useLibrary 'org.apache.http.legacy'
 
     defaultConfig {
         applicationId "com.game.colibri"
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 16
+        targetSdkVersion 29
     }
 
     buildTypes {
         release {
+            // Enables code shrinking, obfuscation, and optimization for only your project's release build type.
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
+            // Enables resource shrinking, which is performed by the Android Gradle plugin.
+            shrinkResources false
+            // Includes the default ProGuard rules files that are packaged with the Android Gradle plugin.
+            proguardFiles getDefaultProguardFile(
+                    'proguard-android.txt'),
+                    'proguard-rules.pro'
         }
     }
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.google.firebase:firebase-core:16.0.1'
-    implementation 'com.google.firebase:firebase-messaging:17.3.4'
-    implementation 'com.google.android.gms:play-services-auth:16.0.1'
-    implementation 'com.google.code.gson:gson:2.2.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.firebase:firebase-core:17.4.4'
+    implementation 'com.google.firebase:firebase-messaging:20.2.4'
+    implementation 'com.google.firebase:firebase-analytics:17.4.4'
+    implementation 'com.google.android.gms:play-services-auth:18.1.0'
+    implementation 'com.google.android.gms:play-services-ads:19.3.0'
+    implementation 'com.google.code.gson:gson:2.8.2'
     implementation 'com.loopj.android:android-async-http:1.4.9'
 }
